requirements.txt
asgiref==3.9.2
Django==5.2.7
djangorestframework==3.14.0
pillow==11.3.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
sqlparse==0.5.3
typing_extensions==4.15.0
